+++ liteidex/src/plugins/litedebug/litedebug.h
@@ -30,6 +30,7 @@
 #include "litebuildapi/litebuildapi.h"
 #include "liteeditorapi/liteeditorapi.h"
 #include "textoutput/textoutput.h"
+#include "processex/processex.h"
 
 class DebugManager;
 class DebugWidget;
@@ -65,6 +66,7 @@ public slots:
     void currentEditorChanged(LiteApi::IEditor*);
     void startDebugExternal();
     virtual void startDebug();
+    virtual void startDebugTests();
     virtual void continueRun();
     virtual void runToLine();
     virtual void stopDebug();
@@ -96,6 +98,7 @@ protected:
     QAction      *m_outputAct;
     QAction *m_startDebugExternal;
     QAction *m_startDebugAct;
+    QAction *m_startDebugTestAct;
     QAction *m_stopDebugAct;
     QAction *m_showLineAct;
     QAction *m_stepIntoAct;