+++ liteidex/deploy/welcome/changes.md
@@ -3,9 +3,17 @@
 ## LiteIDE X Version
   LiteIDE is a simple, open source, cross-platform Go IDE.
 
-### Ver next
-* add litetty plugin
-* gdb enable --tty mode
+### Ver x19
+* MacOS X 10.8 Retina support
+* add simple tty plugin
+* gdb debugger enable --tty mode (windows default)
+* add debug test, thanks for GoEddie <https://github.com/GoEddie>
+* update macosx tool window short key to Ctrl+Alt+1~9
+* markdown preview to fast sync
+* liteeditor add wheel zooming option 
+* fix editor extra width
+* fix init editor update
+* fix find hide
 
 ### 2013.5.21 Ver x18.2
 * fast for gocode